#1363 refactor transaction.xa.manager.atomikos => transaction.xa.manager.
