+++ sharding-transaction/sharding-transaction-xa/src/test/resources/META-INF/services/io.shardingsphere.transaction.manager.xa.XATransactionManager
@@ -1,2 +1,2 @@
-io.shardingsphere.transaction.xa.manager.atomikos.AtomikosTransactionManager
+io.shardingsphere.transaction.xa.manager.AtomikosTransactionManager
